+++ language/src/main/java/i4gl/nodes/sql/DatabaseNode.java
@@ -4,9 +4,7 @@ import com.oracle.truffle.api.frame.FrameSlot;
 import com.oracle.truffle.api.frame.VirtualFrame;
 
 import i4gl.nodes.statement.StatementNode;
-import i4gl.runtime.context.Context;
 import i4gl.runtime.values.Database;
-import i4gl.runtime.values.Sqlca;
 
 public class DatabaseNode extends StatementNode {
     private final FrameSlot slot;
@@ -23,8 +21,7 @@ public class DatabaseNode extends StatementNode {
 
     @Override
     public void executeVoid(VirtualFrame frame) {
-        Sqlca sqlca = Context.get(this).getSqlcaGlobalVariable();
         frame.setObject(slot, database);
-        database.connect(sqlca);
+        database.connect();
     }
 }