+++ language/src/main/java/i4gl/nodes/sql/SqlNode.java
@@ -5,11 +5,9 @@ import com.oracle.truffle.api.interop.InteropLibrary;
 
 import i4gl.nodes.expression.ExpressionNode;
 import i4gl.nodes.statement.StatementNode;
-import i4gl.runtime.context.Context;
 import i4gl.runtime.database.SquirrelExecuterHandler;
 import i4gl.runtime.database.SquirrelSession;
 import i4gl.runtime.values.Database;
-import i4gl.runtime.values.Sqlca;
 import net.sourceforge.squirrel_sql.client.session.SQLExecuterTask;
 
 public class SqlNode extends StatementNode {
@@ -27,9 +25,8 @@ public class SqlNode extends StatementNode {
 
     @Override
     public void executeVoid(VirtualFrame frame) {
-        Sqlca sqlca = Context.get(this).getSqlcaGlobalVariable();
         final Database database = (Database) databaseVariableNode.executeGeneric(frame);
-        database.connect(sqlca);
+        database.connect();
         SquirrelSession session = database.getSession();
         SquirrelExecuterHandler sqlExecuterHandlerProxy = new SquirrelExecuterHandler(session);
         SQLExecuterTask sqlExecuterTask = new SQLExecuterTask(session, sql, sqlExecuterHandlerProxy);